removed code for firefox versions older than 63, since shadow DOM is only active by default from then on;
NOTES:

This means that versions of Firefox older than 63 will not receive this or other future SSS updates. They should be able to keep using version 3.33.1, which is quite stable.

This is a choice I made due to shadow DOM making things much simpler and solving problems on certain webpages. Keeping compatibility would require maintaining a lot of extra code that pre-63 Firefox versions would have to use and defeat most of the purpose of updating to use shadow DOM. Additionally, having to test new updates on multiple Firefox versions and remembering all the different things was becoming a programmer's nightmare, so sticking with 63+ will help.